long to int64_t, in enough places that install succeeds

+++ linbox/linbox/algorithms/matpoly-mult.h
@@ -1058,7 +1058,7 @@ namespace LinBox
 			// find a pseudo nth primitive root of unity
 			for (;;) {
 				// compute the nth primitive root
-				w=  (long) Givaro::powmod(_gen, _pl>>lpts, _p);
+				w=  (int64_t) Givaro::powmod(_gen, int64_t(_pl>>lpts), _p);
 				if ((w !=1) && (w != _pl-1))
 					break;
 
@@ -1278,7 +1278,7 @@ namespace LinBox
 			for (;;) {
 
 				// compute the nth primitive root
-				w=  (long) Givaro::powmod(_gen, _pl>>lpts, _p);
+				w=  (int64_t) Givaro::powmod(_gen, int64_t(_pl>>lpts), _p);
 				//std::cout<<w<<" : "<<_gen<<"\n"<<(_pl>>lpts)<<"\n";
 
 				if ((w !=1) && (w != _pl-1))
@@ -1437,7 +1437,7 @@ namespace LinBox
 			for (;;) {
 
 				// compute the nth primitive root
-				w=  (long) Givaro::powmod(_gen, _pl>>lpts, _p);
+				w=  (int64_t) Givaro::powmod(_gen, int64_t(_pl>>lpts), _p);
 				//std::cout<<w<<" : "<<_gen<<"\n"<<(_pl>>lpts)<<"\n";
 
 				if ((w !=1) && (w != _pl-1))
